Made fixes to medication tab.
Fixed bug that prevented drug interaction tableview from displaying data.

Moved call to saveData from refreshMedicationTables to the button
handlers, to prevent data set from being saved every time a new
character was entered in the add medication text field.

Refactored way medications are selected when trying to view interactions.
If two medications are selected from a tableview those will be the two
to get selected to show interactions. If both tableviews have two drugs
selected the current medication tableview will be selected. This is still
not perfect but it is better than the old way.

#story[746]