Bug fixes for story 26.
Users logged in as a profile can no longer add or delete procedures.
The detailed procedure/edit view can now be opened.
Procedures with the current date are now added to the pending table,
rather than not appearing in either table.
Also fixed the order of procedures, the pending table was not being
ordered and the past table was being ordered incorrectly.

#story[981]